api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "Gert Verhoog",email = "me@gertalot.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "sqlalchemy (>=2.0.0,<3.0.0)",
    "pyyaml (>=6.0.0,<7.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "authlib (>=1.6.0,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)"
]

[tool.poetry]
packages = [
    {include = "app", from = "src"},
    {include = "cli", from = "src"}
]

[tool.poetry.scripts]
savour-admin = "cli.admin:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
